Added binary support to tf, moved pytorch test out of PGD test and into torch test
