+++ tests/attacks/test_projected_gradient_descent.py
@@ -734,66 +734,7 @@ class TestPGD(TestBase):
             np.mean(x_test_adv_np - self.x_test_mnist), np.mean(x_test_adv_fw - self.x_test_mnist), places=6
         )
 
-    def test_pytorch_binary(self):
-        '''
-        This test instantiates a binary classification Pytorch model, then attacks it using PGD
 
-        '''
-        import torch
-        import torch.nn.functional as F
-        import torch.nn as nn
-        import torch.optim as optim
-        import art.estimators.classification
-        import sklearn.datasets
-
-
-        class BasicModel(nn.Module):
-            def __init__(self):
-                super(BasicModel, self).__init__()
-                self.layer_1 = nn.Linear(20, 32)
-                self.layer_2 = nn.Linear(32,1)
-            def forward(self, x):
-                x = F.relu(self.layer_1(x))
-                x= torch.sigmoid(self.layer_2(x))
-
-                return x
-
-
-        device = "cuda"
-        x, y = sklearn.datasets.make_classification(n_samples=10000, n_features=20, n_informative=5, n_redundant=2,
-                                                    n_repeated=0, n_classes=2)
-
-        train_x, test_x, train_y, test_y = sklearn.model_selection.train_test_split(x,y, test_size=0.2)
-        rand_inp = torch.randn((1, 20))
-        check_d_type = type(train_x)
-        model = BasicModel()
-        outs = model(rand_inp)
-        loss_func = nn.BCELoss()
-        model.to(device)
-        opt = optim.Adam(model.parameters(), lr=0.001)
-
-        classifier = art.estimators.classification.PyTorchClassifier(
-            model=model,
-            loss=loss_func,
-            optimizer=opt,
-            input_shape=(1, 28, 28),
-            nb_classes=2,
-        )
-        
-        type_of = type(x)
-
-        test_x_batch = test_x[0:16]
-        test_y_batch = test_y[0:16] 
-        test_loss = classifier.compute_loss(test_x_batch, test_y_batch)
-
-        preds = classifier.predict(test_x_batch)
-        grads = classifier.loss_gradient(test_x_batch, test_y_batch)
-
-        attacker = art.attacks.evasion.ProjectedGradientDescent(classifier)
-
-        generated = attacker.generate(test_x_batch)
-        
-        self.assertIsNot(generated, None)
 
 
 if __name__ == "__main__":
