+++ tests/classifiersFrameworks/test_tensorflow.py
@@ -215,3 +215,39 @@ def test_fgsm_defences(art_warning, fix_get_mnist_subset, image_dl_estimator):
         backend_test_defended_images(attack, fix_get_mnist_subset)
     except ARTTestException as e:
         art_warning(e)
+
+@pytest.mark.only_with_platform("tensorflow2")
+def test_binary_instantiation_and_attack(art_warning):
+    from art.attacks.evasion import ProjectedGradientDescent
+    import art.estimators.classification
+    import art
+    import sklearn
+    import sklearn.datasets
+    import sklearn.model_selection
+
+    tf.compat.v1.disable_eager_execution()
+    try:
+        x, y = sklearn.datasets.make_classification(n_samples=10000, n_features=20, n_informative=5, n_redundant=2,
+                                                    n_repeated=0, n_classes=2)
+        train_x, test_x, train_y, test_y = sklearn.model_selection.train_test_split(x,y, test_size=0.2)
+
+
+        model = tf.keras.models.Sequential([ 
+                                        tf.keras.layers.Dense(128, activation=tf.nn.relu, input_shape = (20, )), 
+                                        tf.keras.layers.Dense(1, activation=tf.nn.sigmoid)
+                                        ])
+        model.summary()
+
+        model.compile(optimizer = tf.optimizers.Adam(),
+                    loss = 'binary_crossentropy',
+                    metrics=['accuracy'])
+
+        classifier = art.estimators.classification.KerasClassifier(model=model)
+        classifier.fit(train_x, train_y, nb_epochs=5)
+        pred = classifier.predict(test_x)
+        attack = ProjectedGradientDescent(estimator=classifier, eps=0.5)
+        x_test_adv = attack.generate(x=test_x)
+        adv_predictions = classifier.predict(x_test_adv)
+        assert adv_predictions is not pred
+    except ARTTestException as e:
+        art_warning(e)
